chore: move claim_asset tests to their own file under each runtime
