+++ integration-tests/emulated/tests/assets/asset-hub-polkadot/src/tests/mod.rs
@@ -19,6 +19,7 @@ mod send;
 mod set_xcm_versions;
 mod teleport;
 mod treasury;
+mod claim_assets;
 
 use crate::*;
 emulated_integration_tests_common::include_penpal_create_foreign_asset_on_asset_hub!(
@@ -27,21 +28,3 @@ emulated_integration_tests_common::include_penpal_create_foreign_asset_on_asset_
 	POLKADOT_ED,
 	system_parachains_constants::polkadot::fee::WeightToFee
 );
-
-#[test]
-fn assets_can_be_claimed() {
-	use asset_hub_polkadot_runtime::ExistentialDeposit;
-	use integration_tests_helpers::test_parachain_can_claim_assets;
-	use xcm_executor::traits::DropAssets;
-
-	let amount = ExistentialDeposit::get();
-	let assets: Assets = (Parent, amount).into();
-
-	test_parachain_can_claim_assets!(
-		AssetHubPolkadot,
-		RuntimeCall,
-		NetworkId::Polkadot,
-		assets,
-		amount
-	);
-}
